Feature register character

+++ internal/data/entity/game/user.go
@@ -23,19 +23,19 @@ type User struct {
 
 type Character struct {
 	Id        int64        `db:"id"`
-	UserID    int          `db:"user_id"`
+	UserID    int64        `db:"user_id"`
 	CreatedAt time.Time    `db:"created_at"` // Consider using time.Time
 	UpdatedAt sql.NullTime `db:"updated_at"` // Consider using time.Time
 
 	Name string `db:"name"`
 
-	Level      int `db:"level"`
-	Experience int `db:"experience"`
+	Level      int64 `db:"level"`
+	Experience int64 `db:"experience"`
 
-	Silver int `db:"silver"`
-	Gold   int `db:"gold"`
-	Gems   int `db:"gems"`
-	Coins  int `db:"coins"`
+	Silver int64 `db:"silver"`
+	Gold   int64 `db:"gold"`
+	Gems   int64 `db:"gems"`
+	Coins  int64 `db:"coins"`
 
 	DragonAmulet bool `db:"dragon_amulet"`
 	PVPStatus    bool `db:"pvp_status"`
@@ -43,7 +43,7 @@ type Character struct {
 	Gender  string `db:"gender"`
 	Pronoun string `db:"pronoun"`
 
-	HairID    int    `db:"hair_id"`
+	HairID    int64  `db:"hair_id"`
 	ColorHair string `db:"color_hair"` // hex string
 	ColorSkin string `db:"color_skin"` // hex string
 	ColorBase string `db:"color_base"` // hex string
@@ -51,13 +51,13 @@ type Character struct {
 
 	QuestID int `db:"quest_id"`
 
-	Strength     int `db:"strength"`
-	Dexterity    int `db:"dexterity"`
-	Intelligence int `db:"intelligence"`
-	Luck         int `db:"luck"`
-	Charisma     int `db:"charisma"`
-	Endurance    int `db:"endurance"`
-	Wisdom       int `db:"wisdom"`
+	Strength     int64 `db:"strength"`
+	Dexterity    int64 `db:"dexterity"`
+	Intelligence int64 `db:"intelligence"`
+	Luck         int64 `db:"luck"`
+	Charisma     int64 `db:"charisma"`
+	Endurance    int64 `db:"endurance"`
+	Wisdom       int64 `db:"wisdom"`
 
 	LastDailyQuestDone string `db:"last_daily_quest_done"`
 
@@ -65,9 +65,9 @@ type Character struct {
 	Skills string `db:"skills"`
 	Quests string `db:"quests"`
 
-	RaceID      int `db:"race_id"`
-	ClassID     int `db:"class_id"`
-	BaseClassID int `db:"base_class_id"`
+	RaceID      int64 `db:"race_id"`
+	ClassID     int64 `db:"class_id"`
+	BaseClassID int64 `db:"base_class_id"`
 
 	LastTimeSeen sql.NullTime `db:"last_time_seen"` // Consider using time.Time
 }
